refactor: SessionService layer 분리
