+++ src/main/java/com/waruru/areyouhere/session/service/SessionCommandService.java
@@ -1,0 +1,11 @@
+package com.waruru.areyouhere.session.service;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface SessionCommandService {
+    public void create(Long courseId, String sessionName);
+    public void delete(List<Long> sessionIds);
+    public void deactivate(Long sessionId);
+    public void setStartTime(Long sessionId, LocalDateTime currentTime);
+}
